Move NodeState out of the SPI
This is no longer referenced by other SPI classes.
