+++ presto-main/src/main/java/io/prestosql/memory/ClusterMemoryManager.java
@@ -77,8 +77,8 @@ import static io.prestosql.SystemSessionProperties.getQueryMaxTotalMemory;
 import static io.prestosql.SystemSessionProperties.resourceOvercommit;
 import static io.prestosql.memory.LocalMemoryManager.GENERAL_POOL;
 import static io.prestosql.memory.LocalMemoryManager.RESERVED_POOL;
-import static io.prestosql.spi.NodeState.ACTIVE;
-import static io.prestosql.spi.NodeState.SHUTTING_DOWN;
+import static io.prestosql.metadata.NodeState.ACTIVE;
+import static io.prestosql.metadata.NodeState.SHUTTING_DOWN;
 import static io.prestosql.spi.StandardErrorCode.CLUSTER_OUT_OF_MEMORY;
 import static java.lang.Math.min;
 import static java.lang.String.format;
