+++ presto-main/src/main/java/io/prestosql/metadata/DiscoveryNodeManager.java
@@ -32,7 +32,6 @@ import io.prestosql.connector.system.GlobalSystemConnector;
 import io.prestosql.failureDetector.FailureDetector;
 import io.prestosql.server.InternalCommunicationConfig;
 import io.prestosql.spi.Node;
-import io.prestosql.spi.NodeState;
 import org.weakref.jmx.Managed;
 
 import javax.annotation.PostConstruct;
@@ -58,9 +57,9 @@ import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static com.google.common.collect.Sets.difference;
 import static io.airlift.concurrent.Threads.threadsNamed;
 import static io.airlift.http.client.HttpUriBuilder.uriBuilderFrom;
-import static io.prestosql.spi.NodeState.ACTIVE;
-import static io.prestosql.spi.NodeState.INACTIVE;
-import static io.prestosql.spi.NodeState.SHUTTING_DOWN;
+import static io.prestosql.metadata.NodeState.ACTIVE;
+import static io.prestosql.metadata.NodeState.INACTIVE;
+import static io.prestosql.metadata.NodeState.SHUTTING_DOWN;
 import static java.util.Locale.ENGLISH;
 import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.Executors.newCachedThreadPool;
