+++ presto-main/src/main/java/io/prestosql/metadata/NodeState.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.spi;
+package io.prestosql.metadata;
 
 public enum NodeState
 {
