+++ presto-main/src/main/java/io/prestosql/metadata/RemoteNodeState.java
@@ -22,7 +22,6 @@ import io.airlift.http.client.Request;
 import io.airlift.json.JsonCodec;
 import io.airlift.log.Logger;
 import io.airlift.units.Duration;
-import io.prestosql.spi.NodeState;
 
 import javax.annotation.Nullable;
 import javax.annotation.concurrent.ThreadSafe;
