+++ presto-main/src/main/java/io/prestosql/server/ServerInfoResource.java
@@ -16,7 +16,7 @@ package io.prestosql.server;
 import io.airlift.node.NodeInfo;
 import io.prestosql.client.NodeVersion;
 import io.prestosql.client.ServerInfo;
-import io.prestosql.spi.NodeState;
+import io.prestosql.metadata.NodeState;
 
 import javax.inject.Inject;
 import javax.ws.rs.Consumes;
@@ -33,8 +33,8 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 import static com.google.common.base.Preconditions.checkState;
 import static io.airlift.units.Duration.nanosSince;
-import static io.prestosql.spi.NodeState.ACTIVE;
-import static io.prestosql.spi.NodeState.SHUTTING_DOWN;
+import static io.prestosql.metadata.NodeState.ACTIVE;
+import static io.prestosql.metadata.NodeState.SHUTTING_DOWN;
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
